style: 抽离颜色转换工具函数
